// apps/web/types/HomeCard.ts
export type TrendDataProps = {
  month: string;
  yearMonth: string;
  income: number;
  expense: number;
};

export type UpcomingExtraProps = {
  month: string;
  yearMonth: string;
  extra: number;
};

export type PreviousSavingsProps = {
  month: string;
  yearMonth: string;
  amount: number;
};

export type ExpensePieDataProps = {
  id: string;
  name: string;
  amount: number;
};